Revert "Feat: social login (#236)"
This reverts commit bcae6fbe2cb068667636941afc81fe0e4d8dfa49.
