+++ src/main/java/com/patientpal/backend/auth/controller/AuthenticationApiV1Controller.java
@@ -6,11 +6,17 @@ import com.patientpal.backend.auth.dto.SignUpRequest;
 import com.patientpal.backend.auth.dto.TokenDto;
 import com.patientpal.backend.auth.service.JwtLoginService;
 import com.patientpal.backend.auth.service.JwtTokenService;
-import com.patientpal.backend.auth.service.TokenBlacklistService;
-import com.patientpal.backend.common.exception.AuthenticationException;
+import com.patientpal.backend.common.exception.BusinessException;
+import com.patientpal.backend.common.exception.EntityNotFoundException;
 import com.patientpal.backend.common.exception.ErrorCode;
+import com.patientpal.backend.member.domain.Member;
+import com.patientpal.backend.member.domain.Role;
 import com.patientpal.backend.member.service.MemberService;
 import com.patientpal.backend.security.jwt.JwtTokenProvider;
+import com.patientpal.backend.security.oauth.CustomOauth2UserPrincipal;
+import com.patientpal.backend.security.oauth.dto.Oauth2SignUpRequest;
+import com.patientpal.backend.security.oauth.dto.Oauth2SignUpResponse;
+import com.patientpal.backend.security.oauth.userinfo.CustomOauth2UserInfo;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.headers.Header;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -20,12 +26,18 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import java.net.URI;
+import java.util.Collections;
+import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.logout.SecurityContextLogoutHandler;
 import org.springframework.web.bind.annotation.CookieValue;
@@ -34,6 +46,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import static com.patientpal.backend.common.exception.ErrorCode.INVALID_USERNAME;
+import static com.patientpal.backend.common.exception.ErrorCode.MEMBER_NOT_EXIST;
 
 @Tag(name = "인증", description = "인증 API")
 @RestController
@@ -44,7 +59,6 @@ public class AuthenticationApiV1Controller {
     private final JwtTokenService tokenService;
     private final JwtTokenProvider jwtTokenProvider;
     private final MemberService memberService;
-    private final TokenBlacklistService tokenBlacklistService;
 
     @PostMapping("/login")
     @Operation(summary = "로그인", description = "로그인을 한다.")
@@ -62,13 +76,7 @@ public class AuthenticationApiV1Controller {
     @ApiResponse(responseCode = "200", description = "재발급 성공",
             content = @Content(schema = @Schema(implementation = AccessTokenResponse.class)),
             headers = @Header(name = "Set-Cookie", description = "리프레시 토큰 쿠키 설정", schema = @Schema(type = "string")))
-    public ResponseEntity<AccessTokenResponse> refresh(@CookieValue("refresh_token") String refreshToken,
-                                                       HttpServletResponse response) {
-
-        if (tokenBlacklistService.isTokenBlacklisted(refreshToken)) {
-            throw new AuthenticationException(ErrorCode.INVALID_TOKEN);
-        }
-
+    public ResponseEntity<AccessTokenResponse> refresh(@CookieValue("refresh_token") String refreshToken, HttpServletResponse response) {
         TokenDto tokenDto = tokenService.refreshJwtTokens(refreshToken);
         setRefreshTokenCookie(response, tokenDto.refreshToken());
         return ResponseEntity.ok().body(new AccessTokenResponse(tokenDto.accessToken()));
@@ -81,13 +89,41 @@ public class AuthenticationApiV1Controller {
         return ResponseEntity.created(URI.create("/api/v1/members/" + memberService.save(request))).build();
     }
 
+    @GetMapping("/oauth2/register")
+    @Operation(summary = "소셜 회원가입 정보 가져오기", description = "소셜 로그인 후 회원가입 페이지에서 필요한 정보를 가져온다.")
+    @ApiResponse(responseCode = "200", description = "소셜 회원가입 정보 가져오기 성공",
+            content = @Content(schema = @Schema(implementation = Oauth2SignUpResponse.class)))
+    public ResponseEntity<Oauth2SignUpResponse> getOauth2UserInfo(HttpSession session) {
+        Oauth2SignUpResponse response = createOauth2SignUpResponse(session);
+        return ResponseEntity.ok(response);
+    }
+
+    @PostMapping("/oauth2/register")
+    @Operation(summary = "소셜 회원가입", description = "소셜 로그인 후 추가 정보를 입력받아 회원가입을 한다.")
+    @ApiResponse(responseCode = "201", description = "회원가입 성공",
+            content = @Content(schema = @Schema(implementation = Oauth2SignUpResponse.class)))
+    public ResponseEntity<Oauth2SignUpResponse> processOauth2Signup(@Valid @RequestBody Oauth2SignUpRequest signupForm,
+                                                                    HttpSession session) {
+        String username = getUsername(signupForm, session);
+
+        memberService.saveSocialUser(signupForm);
+
+        Member member = getMember(username);
+        String token = generateToken(member, signupForm.getProvider(), signupForm.getEmail(), signupForm.getName());
+
+        session.setAttribute("token", token);
+
+        Oauth2SignUpResponse response = Oauth2SignUpResponse.fromMember(member, signupForm.getEmail(), signupForm.getName(), signupForm.getRole().name(), signupForm.getProvider(), token);
+
+        return ResponseEntity.created(URI.create("/api/v1/members/" + member.getId())).body(response);
+    }
+
     @GetMapping("/logout")
     @Operation(summary = "로그아웃", description = "사용자를 로그아웃한다.")
     public void logout(HttpServletRequest req, HttpServletResponse resp,
                          @CookieValue("refresh_token") String refreshToken) {
         if (refreshToken != null) {
             jwtTokenProvider.invalidateToken(refreshToken);
-            setRefreshTokenCookie(resp);
         }
         new SecurityContextLogoutHandler()
                 .logout(req, resp, SecurityContextHolder.getContext().getAuthentication());
@@ -105,12 +141,52 @@ public class AuthenticationApiV1Controller {
         response.addHeader(HttpHeaders.SET_COOKIE, refreshTokenCookie.toString());
     }
 
-    private static void setRefreshTokenCookie(HttpServletResponse resp) {
-        Cookie cookie = new Cookie("refresh_token", null);
-        cookie.setPath("/");
-        cookie.setHttpOnly(true);
-        cookie.setMaxAge(0);
-        resp.addCookie(cookie);
+    private Oauth2SignUpResponse createOauth2SignUpResponse(HttpSession session) {
+        String email = (String) session.getAttribute("email");
+        String name = (String) session.getAttribute("name");
+        Role role = (Role) session.getAttribute("role");
+        String provider = (String) session.getAttribute("provider");
+        String username = (String) session.getAttribute("username");
+        String token = (String) session.getAttribute("token");
+
+        if (username == null || username.isEmpty()) {
+            throw new BusinessException(INVALID_USERNAME, "Username is not set in session");
+        }
+
+        Member member = memberService.getUserByUsername(username);
+
+        if (member == null) {
+            throw new EntityNotFoundException(MEMBER_NOT_EXIST, "Member not found with username: " + username);
+        }
+
+        return Oauth2SignUpResponse.fromMember(member, email, name, role.name(), provider, token);
+    }
+
+    private String generateToken(Member member, String provider, String email, String name) {
+        CustomOauth2UserInfo userInfo = CustomOauth2UserInfo.of(provider, Map.of("email", email, "name", name));
+        String password = member.getPassword() != null ? member.getPassword() : "DUMMY_PASSWORD";
+        CustomOauth2UserPrincipal principal = new CustomOauth2UserPrincipal(member.getUsername(), password, Collections.singletonList(new SimpleGrantedAuthority(member.getRole().name())));
+        principal.setUserInfo(userInfo);
+
+        Authentication authentication = new UsernamePasswordAuthenticationToken(principal, null, principal.getAuthorities());
+        return jwtTokenProvider.createAccessToken(authentication);
+    }
+
+    private String getUsername(Oauth2SignUpRequest signupForm, HttpSession session) {
+        String username = (String) session.getAttribute("username");
+        signupForm
+                .toBuilder()
+                .username(username)
+                .build();
+        return username;
+    }
+
+    private Member getMember(String username) {
+        Member member = memberService.getUserByUsername(username);
+        if (member == null) {
+            throw new BusinessException(ErrorCode.MEMBER_NOT_EXIST, "Member not found with username: " + username);
+        }
+        return member;
     }
 }
