+++ src/main/resources/application.yml
@@ -80,7 +80,6 @@ spring:
   config:
     activate:
       on-profile: "local"
-      include: oauth2
 
   datasource:
     driver-class-name: org.h2.Driver
@@ -112,7 +111,6 @@ spring:
     execution:
       pool:
         max-size: 16
-
 
 ---
 spring:
