--- src/test/java/com/patientpal/backend/fixtures/auth/SocialDataFixture.java
@@ -1,13 +1,0 @@
-package com.patientpal.backend.fixtures.auth;
-
-public class SocialDataFixture {
-
-    public static final String MOCK_ACCESS_TOKEN = "mockAccessToken";
-    public static final String GOOGLE_PROVIDER = "google";
-    public static final String KAKAO_PROVIDER = "kakao";
-    public static final String NAVER_PROVIDER = "naver";
-
-    public static final String GOOGLE_USER_DATA = "{\"email\": \"test@example.com\", \"name\": \"Test User\"}";
-    public static final String KAKAO_USER_DATA = "{\"email\": \"test@example.com\", \"nickname\": \"Test User\"}";
-    public static final String NAVER_USER_DATA = "{\"email\": \"test@example.com\", \"name\": \"Test User\"}";
-}
