+++ src/test/java/com/patientpal/backend/fixtures/member/MemberFixture.java
@@ -11,7 +11,6 @@ public class MemberFixture {
     public static final String DEFAULT_CONTACT = "010-1234-5678";
     public static final String DEFAULT_PASSWORD = "password123";
     public static final Long DEFAULT_ID = 1L;
-    public static final Integer DEFAULT_AGE = 30;
 
     public static Member createDefaultMember() {
         return createMemberBuilder().build();
@@ -35,8 +34,7 @@ public class MemberFixture {
                 .password(DEFAULT_PASSWORD)
                 .provider(DEFAULT_PROVIDER)
                 .role(Role.USER)
-                .contact(DEFAULT_CONTACT)
-                .age(DEFAULT_AGE);
+                .contact(DEFAULT_CONTACT);
     }
 
     private static Member.MemberBuilder createMemberBuilderWithUserRole() {
